# packages/api-client/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Yetano Work API",
    "description": "Public API for Yetano Work.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/cases/{caseId}/status-history": {
      "get": {
        "operationId": "listCaseStatusHistory",
        "description": "Lists immutable status history for one organization-scoped case.",
        "parameters": [
          {
            "in": "path",
            "name": "caseId",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Unique identifier of a case.",
              "format": "uuid",
              "title": "CaseId"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Opaque cursor returned by a previous case-list request.",
              "minLength": 1,
              "title": "CaseCursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Maximum number of cases returned in one page.",
              "maximum": 100,
              "minimum": 1,
              "title": "CaseListLimit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "nextCursor"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "actorId",
                          "actorType",
                          "caseId",
                          "caseVersion",
                          "changedAt",
                          "fromStatus",
                          "id",
                          "note",
                          "source",
                          "toStatus",
                          "transitionId",
                          "type"
                        ],
                        "properties": {
                          "actorId": {
                            "type": "string",
                            "minLength": 1
                          },
                          "actorType": {
                            "anyOf": [
                              {
                                "type": "string",
                                "const": "system"
                              },
                              {
                                "type": "string",
                                "const": "user"
                              }
                            ]
                          },
                          "caseId": {
                            "type": "string",
                            "description": "Unique identifier of a case.",
                            "format": "uuid",
                            "title": "CaseId"
                          },
                          "caseVersion": {
                            "type": "integer",
                            "minimum": 1
                          },
                          "changedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "fromStatus": {
                            "anyOf": [
                              {
                                "anyOf": [
                                  {
                                    "type": "string",
                                    "const": "new"
                                  },
                                  {
                                    "type": "string",
                                    "const": "postponed"
                                  },
                                  {
                                    "type": "string",
                                    "const": "working"
                                  },
                                  {
                                    "type": "string",
                                    "const": "waiting"
                                  },
                                  {
                                    "type": "string",
                                    "const": "resolved"
                                  },
                                  {
                                    "type": "string",
                                    "const": "canceled"
                                  }
                                ],
                                "description": "Current lifecycle status of a case.",
                                "title": "CaseStatus"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "note": {
                            "anyOf": [
                              {
                                "type": "string",
                                "maxLength": 2000,
                                "minLength": 1
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "source": {
                            "anyOf": [
                              {
                                "type": "string",
                                "const": "migration"
                              },
                              {
                                "type": "string",
                                "const": "runtime"
                              }
                            ]
                          },
                          "toStatus": {
                            "anyOf": [
                              {
                                "type": "string",
                                "const": "new"
                              },
                              {
                                "type": "string",
                                "const": "postponed"
                              },
                              {
                                "type": "string",
                                "const": "working"
                              },
                              {
                                "type": "string",
                                "const": "waiting"
                              },
                              {
                                "type": "string",
                                "const": "resolved"
                              },
                              {
                                "type": "string",
                                "const": "canceled"
                              }
                            ],
                            "description": "Current lifecycle status of a case.",
                            "title": "CaseStatus"
                          },
                          "transitionId": {
                            "anyOf": [
                              {
                                "type": "string",
                                "format": "uuid"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "type": {
                            "anyOf": [
                              {
                                "type": "string",
                                "const": "created"
                              },
                              {
                                "type": "string",
                                "const": "transitioned"
                              }
                            ]
                          }
                        },
                        "additionalProperties": false,
                        "description": "An immutable entry in a case status history.",
                        "title": "CaseStatusChange"
                      }
                    },
                    "nextCursor": {
                      "anyOf": [
                        {
                          "type": "string",
                          "minLength": 1
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "additionalProperties": false,
                  "description": "A cursor-paginated case status history.",
                  "title": "CaseStatusHistory"
                }
              }
            },
            "description": "A cursor-paginated case status history."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "title",
                    "type"
                  ],
                  "properties": {
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "maximum": 599,
                      "minimum": 400
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string",
                      "format": "uri-reference"
                    }
                  },
                  "additionalProperties": true,
                  "description": "RFC 9457 problem details returned by the Yetano Work API.",
                  "title": "ProblemDetails"
                }
              }
            },
            "description": "The request is invalid."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "title",
                    "type"
                  ],
                  "properties": {
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "maximum": 599,
                      "minimum": 400
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string",
                      "format": "uri-reference"
                    }
                  },
                  "additionalProperties": true,
                  "description": "RFC 9457 problem details returned by the Yetano Work API.",
                  "title": "ProblemDetails"
                }
              }
            },
            "description": "Authentication is required."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "title",
                    "type"
                  ],
                  "properties": {
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "maximum": 599,
                      "minimum": 400
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string",
                      "format": "uri-reference"
                    }
                  },
                  "additionalProperties": true,
                  "description": "RFC 9457 problem details returned by the Yetano Work API.",
                  "title": "ProblemDetails"
                }
              }
            },
            "description": "The actor lacks a required capability."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "title",
                    "type"
                  ],
                  "properties": {
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "maximum": 599,
                      "minimum": 400
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string",
                      "format": "uri-reference"
                    }
                  },
                  "additionalProperties": true,
                  "description": "RFC 9457 problem details returned by the Yetano Work API.",
                  "title": "ProblemDetails"
                }
              }
            },
            "description": "The case does not exist in the active organization."
          }
        },
        "tags": [
          "Cases"
        ]
      }
    },
...
  },
...
}